Update README.md - Reddit link
The purpose of this small edit ( fr.reddit.com --> np.reddit.com ) is so that the user isn't redirected to the French "version" of Reddit, but to the English, Read-Only (No voting, etc) "version" of the site. np.reddit.com is usually used when being linked to by other subreddits or other sites, such as this.

Thank you very much for your time!