Replace prebake with Custom prebake
I'd like to suggest to replace the original prebake EU cookie filtering with my own because the original has been dead for almost a year. The original author also doesn't reply to mails.

I'm willing to keep this list updated and my version also has indications when the filter was last tested, with the ! LC: date comment, which makes the usefulness of the filter a bit more clear.

EU is deciding to adjust this law where these annoying messages are not necessary anymore, but as long as this is not in place, a solution such as this is necessary to remove these annoyances as much as possible.