Sixth pass to re-design the dashboard
Related commits:
- https://github.com/gorhill/uBlock/commit/0662767dd62731320473e9c96a17077145a00610
- https://github.com/gorhill/uBlock/commit/99b2a0a761eec860132bc08680710c3c785f90b6
- https://github.com/gorhill/uBlock/commit/a8e3118feafd5c2d7c3e3d03727beff17f9928b4
- https://github.com/gorhill/uBlock/commit/453f5450b6e6550618bf97b5b0541f5f46c35d8e
- https://github.com/gorhill/uBlock/commit/a45a21f337216d7d4aaddb5a454ed4842eca3806

This commit focuses on implementing the use of
CSS `var(...)` throughout so as to make it easy
to create themes -- a requirement for uBO to
support a dark theme.

There is still work to do regarding converting
uBO's CSS to completely support `var(...)` but
being able to start using theming will help
complete and fine tune CSS `var(...)` support.

Related issue:
- https://github.com/uBlockOrigin/uBlock-issues/issues/401