Add widget to control selector depth to element picker
Further iterating on the work done in following commit:
- https://github.com/gorhill/uBlock/commit/1268f0ae436b039b9f7f6bc5e61568c4f841179e

This commit adds a new widget to the element picker to
control the depth of a cosmetic filter selector. The
new widget is essentially just another way of selecting
the depth, which is still controllable through picking
one of the cosmetic filters in the list of candidates.